refactor(GUI): ad boilerplate for GTK3 scrolledwindow
