+++ src/ramstk/views/gtk3/widgets/__init__.py
@@ -18,3 +18,4 @@ from .dialog import (
 from .entry import RAMSTKEntry, RAMSTKTextView
 from .frame import RAMSTKFrame
 from .label import RAMSTKLabel, do_make_label_group
+from .scrolledwindow import RAMSTKScrolledWindow
